scraping service: renamed scraper.py to service.py for unification - scraping class object is now newly created when injected in every use case
