+++ infra/scraper/src/di/dependencies.py
@@ -1,6 +1,9 @@
-from scraper.scraper import TrackerGGScraper
-
-scraper = TrackerGGScraper()
+from scraper.service import TrackerGGScraper
 
 def get_scraper():
-    return scraper
+    scraper = TrackerGGScraper()
+
+    yield scraper
+    # anything that needs to be executed after the function that's being injected with this dependency should go here
+    # for example, closing a db connection
+
